feat(release): close ADR-0073 C07 BIGINT cutover

+++ android/app/src/main/java/com/ticketbox/domain/model/Expense.kt
@@ -332,7 +332,7 @@ data class MonthComparison(
     val currentAmountCents: Long,
     val previousAmountCents: Long,
     val deltaAmountCents: Long,
-    val percentChange: Int?,
+    val percentChange: Long?,
 )
 
 data class BudgetProgress(
@@ -341,7 +341,7 @@ data class BudgetProgress(
     val spentCents: Long,
     val remainingCents: Long,
     val progress: Float,
-    val percent: Int,
+    val percent: Long,
     val overBudget: Boolean,
 )
